Move func helpers to Utility.h
